+++ test/result.test.js
@@ -8,9 +8,27 @@ test("renders without crashing", () => {
   let selected = null;
   const onSelect = result => { selected = result; };
 
-  const component = shallow(<Result result={results[0]} onSelect={onSelect} />);
+  const component = shallow(<Result result={results.search[0]} onSelect={onSelect} />);
   expect(component.type()).toEqual("button");
 
   component.simulate("click");
-  expect(selected).toEqual(results[0]);
+  expect(selected).toEqual(results.search[0]);
+});
+
+test("loads the image in the background", () => {
+  const component = shallow(<Result result={results.search[0]} />);
+  expect(component.find("span")).toHaveLength(0);
+
+  component.instance().image.onload();
+  component.update();
+
+  expect(component.find("span")).toHaveLength(1);
+});
+
+test("does not attempt to set state if the image finishes after unmount", () => {
+  const component = shallow(<Result result={results.search[0]} />);
+  const { image } = component.instance();
+
+  component.unmount();
+  image.onload();
 });