+++ test/suggestion.test.js
@@ -1,0 +1,18 @@
+import React from "react";
+import { shallow } from "enzyme";
+
+import Suggestion from "../src/suggestion";
+
+test("renders without crashing", () => {
+  let selected = null;
+  const onSuggestionClick = result => { selected = result; };
+
+  const component = shallow(
+    <Suggestion suggestion="test" onSuggestionClick={onSuggestionClick} />
+  );
+
+  expect(component.type()).toEqual("button");
+
+  component.simulate("click");
+  expect(selected).toEqual("test");
+});