+++ test/test-server.js
@@ -1,28 +1,51 @@
 import http from "http";
 
-export const results = [
-  { id: 1, media: [{ tinygif: { url: "https://via.placeholder.com/10x10" } }] },
-  { id: 2, media: [{ tinygif: { url: "https://via.placeholder.com/10x10" } }] },
-  { id: 3, media: [{ tinygif: { url: "https://via.placeholder.com/10x10" } }] },
-  { id: 4, media: [{ tinygif: { url: "https://via.placeholder.com/10x10" } }] },
-  { id: 5, media: [{ tinygif: { url: "https://via.placeholder.com/10x10" } }] }
-];
+let counter = 0;
 
-const withTestServer = async (port, callback) => {
+const makeId = () => {
+  counter += 1;
+  return counter;
+};
+
+const makeResult = () => ({
+  id: makeId(),
+  itemurl: "https://tenor.com/view/this-is-a-test-gif-12345",
+  media: [{ tinygif: { url: "https://via.placeholder.com/10x10" } }]
+});
+
+export const results = {
+  autocomplete: ["test", "testing", "test2", "testingtesting", "testy testerson"],
+  search_suggestions: ["test", "unit test", "acceptance test", "testing", "how to test"],
+  search: [makeResult(), makeResult(), makeResult(), makeResult(), makeResult()]
+};
+
+const getRequestKey = url => (
+  url.slice(1).substring(0, url.indexOf("?") - 1)
+);
+
+const withTestServer = (port, callback) => async () => {
   const server = http.createServer();
-  server.requests = 0;
+
+  server.requests = Object.keys(results).reduce((accum, key) => (
+    Object.assign({}, accum, { [key]: 0 })
+  ), {});
 
   server.on("request", (request, response) => {
-    server.requests += 1;
+    const requestKey = getRequestKey(request.url);
+    server.requests[requestKey] += 1;
+
     response.writeHead(200, { "Content-Type": "application/json" });
-    response.write(JSON.stringify({ results }));
+    response.write(JSON.stringify({ results: results[requestKey], next: "12" }));
     response.end();
   });
 
   server.listen(port);
-  await callback(server);
 
-  server.close();
+  try {
+    await callback(server);
+  } finally {
+    server.close();
+  }
 };
 
 export default withTestServer;