+++ CHANGELOG.md
@@ -6,6 +6,19 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+## [1.1.0] - 2018-10-03
+### Added
+- Type ahead through the use of Tenor's autocomplete feature.
+- Suggestions are now rendered through the user of Tenor's suggestions feature.
+- Pagination is now supported. Additionally you can use the meta-key plus left arrow and right arrow to go between pages.
+
+### Changed
+- Now encoding the URI being sent to tenor to ensure it's a valid URL.
+- If you click outside the component, the component now knows about that click and closes the selector.
+- The search bar is now of type "search" which builds in some nicities from the browsers that support it.
+- Added better accessibility support by properly naming the GIF buttons.
+- Load preview GIFs in the background and then set them to fade in once they are loaded.
+
 ## [1.0.0] - 2018-09-26
 ### Added
 - The optional `contentRef` prop to get access to the actual div that is being rendered.